Trying to get the event listener inheritance working
Without success. There seems to be missing some key concepts or some
pointers to the base class. Is there mixed inheritance (JS->CLR,
CLR->JS) planned? So why don't we let the CLR code inheritance do the
work, we would just have to find the CLR defining class and bind to that
method?